Final attempt to continue using my system, going to recode the handling of all of this
Attempted to get my old system of getting things to work with more that 1 client working but to no avail.
I plan to recode it so that it is much easier to understand how the steps are possessed and stop trying to use math to automatically calculate who's turn and the next expected step as having to add extra checks for a reason shouldn't happen, the single function call should do it all.