Fixed small problems, started but then undid a change
Using the wrong variable name for thread joining
Going to stay away from using 3.8 python things (dolphin operator on 771)
I had started making a builtin to decide if the theoretical model should be loaded or a real world but then I decided that once I get this theoretical model made, I will put it on a seperate branch.
I will then continue work on the real-world model on master instead of trying to keep them in the same project and to keep the project size small when loading on devices.