Got searching working and a good printout
Added map attribute which just is a shortcut to call search.
Added serach to do the recursive search through folders.
Added to string to allow for better printing and I changed repr to use this.
Repr just calls __str__ to a get a string version of the object.
ROOT object's folder was changed to the data folder that doesn't appear in commits as it is empty and is generated at runtime (the files within are generated at runtime, not the folder itself).
I plan to add in the File object so that I don't have to keep using a None as a placeholder for it in searching.